requirements.txt
flask
pandas
sqlalchemy
pymysql
pyodbc
azure-storage-blob
transformers
torch
scikit-learn
openpyxl
